Removed the UUID from IO Switchboard tests.
The IO switchboard server creates a UNIX socket at a given path.
Due to OS constraints, this path must be less than 104 characters long.

In the tests, the path is set to a value based on the test directory.
If the test directory is too long, the UNIX socket creation will fail,
as observed in OSX, where the standard temporary directory does not
default to `/tmp` (as is the case on most Linux's).

The test directory was changed to provide platform-specific values
in this review:
https://reviews.apache.org/r/54490/

This commit shortens the UNIX socket address by removing the UUID.
This is safe because we are not running multiple IO switchboards
in the same test, in the same directory.