Acked correct task id in DefaultExecutorTest.KillTask test.
When acknowledging task status updates, this test was erroneously
using a cached task id instead of the actual task id sent in the
status update. This can lead to failed acknowledgements when
status updates appear in different order.

This patch just uses the task id sent as part of the status update in
the acknowledgement.

Review: https://reviews.apache.org/r/54663/