Windows: Fix build break in libprocess tests.
Commit f61f321b calls a macro, `INSTANTIATE_TEST_CASE_P`, with some
arguments `#ifdef`'d. This breaks MSVC 1900 build; the specification
does not require the text a macro expands to be processed, which means
that the un-processed `#ifdef` ends up in the expansion of
`INSTANTIATE_TEST_CASE_P`, which causes a build-time error.

This commit will resolve this problem by moving the `#ifdef` to surround
the invocation of `INSTANTIATE_TEST_CASE_P` instead of being around the
arguments to the macro.

Review: https://reviews.apache.org/r/54685/