Used correct shell string comparison operator in test.
In order to compare two strings for equality in shell with the `[`
operator (which is often mapped to `test`), the `=` operator should be
used.

This patches replaces usage of `==` with `=` in one such comparison.
While on many system this appears to work with the shell mapped to
`sh`, it does not work for e.g., ubuntu which uses `dash` as the
default shell.

Review: https://reviews.apache.org/r/54739/