Fixed the 'IOSwitchboardTest.KillSwitchboardContainerDestroyed' test.
The container was launched with TTY enabled. This meant that
killing the switchboard would trigger the task to terminate
on its own owing to the "master" end of the TTY dying. This
would make it not go through the code path of the isolator
failing due to resource limit issue.

Review: https://reviews.apache.org/r/54770