Stout: Added helper for determining a runtime directory.
This is a continuation of the fix in:
https://reviews.apache.org/r/54336/

The new helper `os::var()` returns `/var` on POSIX and (usually)
`C:\\ProgramData` on Windows.  On Windows, this uses a COM API to look
up the correct location for persistent, app-local (but not per user)
variable data.

The correct place for variable runtime data on Windows is not
necessarily in a temporary directory, but in the analogous location
`ProgramData`.  Thus we need a platform-agnostic way to refer to `/var`.

NOTE: The call to `ShGetKnownFolder` is not RAII because it is a C API,
and the ATL `CComHeapPtr` class is not used in this commit due to
Windows header issues.  Thus the buffer allocated by the C API is freed
immediately after the data is copied into a `std::wstring`.  Because the
Windows API returns a UTF-16 string, and Unicode characters are valid
in Windows path names, we have to correctly convert it to UTF-8 using
`<codecvt>`.

Review: https://reviews.apache.org/r/54335/