Used `os::var()` for the default value in agent `--runtime_dir`.
This commit improves the logic of `Flags::runtime_dir` to use
`os::var()` for encapsulation of `/var` vs `C:\\ProgramData`.

The flags still contain some platform specific `#ifdef`-ing as the
desired runtime path is structured differently.  On POSIX, the
default is `/var/run/mesos`.  On Windows, the default is
`C:\\ProgramData\mesos\runtime`.  Notice that the relative position of
"run(time)" is flipped.

The permission check for the default path is fixed to check actual
access permissions to the directory, instead of assuming permissions
based on comparison of usernames.

Stylistically we chose a bit of duplication as a trade-off for enhanced
readability and for auto-formatting to work properly.

Review: https://reviews.apache.org/r/54515/