Added a fixed-capacity hashmap to stout.
This is similar to the LinkedHashMap and Cache constructs.

Unlike LinkedHashMap, a BoundedHashMap has a fixed maximum capacity:
when the BoundedHashMap is at capacity, inserting another element
results in removing the oldest element still in the map (i.e., FIFO
order is used for eviction).

Unlike Cache, reading an element from a BoundedHashMap does not change
which entry is evicted next.

Review: https://reviews.apache.org/r/54178/