Added some missing log statements for scheduler calls.
Some Scheduler API calls ('KILL'/'SHUTDOWN') were missing info
level logging statements. In the case of `KILL', the logging
statement was there but was invoked in `killTask()` but was
not consistent with how we were logging for the other calls
in the continuation helper.

Review: https://reviews.apache.org/r/54913/