Removed dependency of `recover` on cached networks.
Since CNI networks can be added and removed on the fly, and for a
given container the CNI configuration with which the container was
launched is check-pointed in the containers \`network/cni\` work
directory, it doesn't make sense to require the network name that the
container is associated with to exist in the in-memory cache. Hence,
removing these dependencies.

Review: https://reviews.apache.org/r/54718/