Leaked module libraries to avoid inconsistencies during unloading.
The module manager holds a `static` map of RAII handles to loaded
dynamic libraries. The handles, e.g., make sure to `dclose` loaded
libraries when the handle is destructed.

Since the handle is a `static` variable, it will only be destructed
when the library gets unloaded. This might lead to inconsistencies
when libmesos's own destruction unloads e.g., a dynamic libprocess
also opened as part of a dynamic libraries `dlopen`, as the module's
cleanup would not find libprocess anymore.

Instead leak the handles to the `dlopen`'ed modules to "clean them up
implicitly" by the unloading.

Review: https://reviews.apache.org/r/54753/