Fixed more tests to pass when `HAS_AUTHENTICATION` is undefined.
This is a follow-on for https://reviews.apache.org/r/54803/

Currently, when `HAS_AUTHENTICATION` is undefined, the Agent will
use `delay` to schedule a random time in the future to register with the
Master, to avoid the thundering herd problem after a Master failover.
The authentication codepath, in contrast, schedules the registration
immediately.

In tests where we have `Clock::pause`'d when we are supposed to be
registering the slave, the authention codepath will succeeed, while
no-authentication codepath will hang forever.

A much more detailed analysis of this situation exists in MESOS-6803.

This commit will resolve this issue for a few existing test files by
changing the tests to advance/settle the clock as necessary.

Review: https://reviews.apache.org/r/54910/