Added initial random delay to agent (re)registration.
Currently when a master fails over, agents that do not use
authentication will choose a random time to re-register to avoid the
"thundering herd" problem. This is not true of the authenticated
codepath -- agents that have a credential will attempt to re-register
immediately.

This issue adds a random delay to the initial authentication to match
the non-authenticated code path.

This issue caps off a chain of reviews that fixes tests to work with
this change. This issue also resolves MESOS-6803.

Review: https://reviews.apache.org/r/54928/