Handled all possible offers in FaultToleranceTest.FrameworkReregister.
The test FaultToleranceTest.FrameworkReregister observes a single
offer in order to monitor progress and as a trigger to initiate other
actions later in the test. This change installs expectations for
possible additional offers. This allows for the thread running the
test body to be delayed with respect to the master thread which could
in the meantime make additional offers.

Review: https://reviews.apache.org/r/54889/