Fixed copy-template-and-create-symlink make target.
This target was `PHONY`, but not declared as such. It also explicitly
used `ln` with a `--force` flag instead of the more generally
applicable and suggested `$(LN_S)` we use elsewhere, and was
inconsistent on whether content was linked or copied.

This change fixes above issues. We also use independent shell commands
for independent commands and fix some whitespace issues.

Review: https://reviews.apache.org/r/54896/