Made `slaves.recovered` in master a hashmap.
This information will be used in both the
\`GetAgents\` API call and \`/state\` endpoint
to return \`AgentInfo\` of recovered but unregistered
slave, to facilitate operator to better handle
such cases.

The reason we want to return \`AgentInfo\` instead
of \`AgentID\` is that hostname and ip address
can be found in former, but not latter.

Review: https://reviews.apache.org/r/53095/