CMake: Removed special case for `BUILD_DIR` on Windows.
Currently in the test harness, we pass the flags `-DSOURCE_DIR` and
`-DBUILD_DIR` to the compiler. These are used to do things like
determine where the containerizer binary is to launch it for tests.

On Windows, these paths should be Windows-style, with '\' characters
separating path components. Unfortunately, CMake does not escape these
slashes in path strings, so when we pass them as preprocessor flags,
a string like `C:\src` will look to the standard Windows API like a
string with an escaped '\s' character.

On the other hand, Windows APIs are happy to take Unix-style paths as
arguments. So, to unblock making the agent tests work, we simply use
Unix paths here.

Review: https://reviews.apache.org/r/55025/