CMake: Added correct values for some build flags.
Currently, the CMake build does not implement certain build flags
as the autotools solution does. For example, `BUILD_DATE` is hard coded,
and the `BUILD_USER` flags is always "frank". This commit will add
correct values for all such flags, with the exception of `BUILD_FLAGS`.

In addition, this commit will take this opportunity to clean up the
organization of the build flags. Currently, we call `add_definitions`
individually for each flag we wish to define. This commit will introduce
a new variable, `MESOS_CPPFLAGS`, which holds the build flags for the
project. We call `add_definitions` once, when `MESOS_CPPFLAGS` is
completely populated.

Review: https://reviews.apache.org/r/55040/