Fix SIGBUS crash on ARM64/AArch64.
Currently in the Linux launcher when the stack is allocated and prepared
for a call to clone() it is not properly aligned. This is not an issue
for x86 or x64 but for ARM64/AArch64 it is because of the requirement of
having the stack aligned to a 16 byte boundary. While x86 and x64 also
expect the stack to have a 16 byte aligned stack, it is not enforced. An
explanation of the stack and requirements for ARM64 can be found here
https://goo.gl/algH07 (specifically section 5.2.2.1 that says SP mod 16
= 0. The stack must be quad-word aligned.).

Additionally, the way that the stack is currently allocated and passed
to clone() accidentally chops off one entry, making a stack overflow
using those missing 8 bytes a possibility. Fixing this while aligning
the memory will fix both the issue of the stack overflow issue as well
as the SIGBUS crash. We should also net better performance from having
the stack aligned.

Review: https://reviews.apache.org/r/54996/