Windows: Stout: Removed dependency on Shell API.
The API `SHGetKnownFolderPath` requires `Shell32.dll`, which is not
available on Nano server (a version of Windows Server 2016).
An equivalent API `GetAllUsersProfileDirectory` is.

This API is also friendlier, as we own the allocation.

The Unicode version `GetAllUsersProfileDirectoryW` is explicitly used
so that we are guaranteed a Unicode path, which we then convert from
UTF-16 to UTF-8, instead of using the ANSI version which depends on a
varying Windows code-page, and is not recommended.

A `vector<wchar_t>` is used over a `wstring` to avoid dealing
with the placement of the null-terminating character.

Review: https://reviews.apache.org/r/54877/