Windows: Modified `argv[0]` of mesos-containerizer launch command.
On POSIX, the value of `argv[0]` generally makes no difference when
launching `mesos-containerizer`.  On Windows however, we will not
be able to find the `mesos-containerizer.exe` binary as we launch
subprocesses from `argv[0]` directly (rather than `exec`-ing a
path with some `argv`).

This fills in the value of `argv[0]` to match the binary path.

Review: https://reviews.apache.org/r/55023/