Windows: Added parent hooks to subprocess.
The subprocess API exposes various hooks to customize behavior as we
create the child process. Currently, these hooks are unimplemented in
the Windows codepaths. This commit will implement the parent hooks --
which are executed after the child process is created, but before it
begins execution -- on the Windows codepath.

This commit implements parents hooks and adds a `ParentHook` for
placing the new subprocess in a Job Object, which will allow us to
kill the subprocess plus any of its children.

There is some subtlety in making sure _all_ processes associated with a
task are captured inside a Job Object. The most important consideration
is that we need to make sure to add any process to the Job Object before
it has a chance to create any child processes; if we fail to do this,
the children will not be captured in the Job Object.

The solution to this is fairly simple on Windows. The process creation
API allows users to trivially create a process in a suspended state, so
that the Windows kernel scheduler does not schedule the process to run
until the user explicitly resumes the main thread. This allows us to
create the process and add it to a Job Object before it has a chance to
create children, and then start the process.

Review: https://reviews.apache.org/r/55312/