Fixed unsafe usage of process pointer in async.hpp.
The `async(...)` helper spawns a libprocess actor to execute some
lambda (asynchronously, of course).  This actor is owned by the
libprocess GC actor, but the `AsyncExecutor` stores a copy of that
pointer and dereferences it in several possible locations.

Instead, the `AsyncExecutor` should store the `PID` of the actor,
which can be safely used, even if the actor has already terminated;
such as during libprocess finalization.

Review: https://reviews.apache.org/r/55701