Fixed expectations of `containerizer->containers()` in tests.
Several tests that intercept and act upon this containerizer method
may cause the test suite to segfault if the `containers()` method
returns an empty result.  This is due to the use of `EXPECT_EQ`
instead of `ASSERT_EQ` when checking the return value.  Subsequent
code sometimes assumes the future contains a non-empty result.