Made `AttachContainerOutput/Input` tests use an existing container ID.
These tests verify that the Agent returns a 500 if the containerizer
doesn't support the `attach` call.

This chain refactors those methods and make them return a 404 instead of
a 500 if the container can't be found, so we need to pass the ID of an
existing container.

Review: https://reviews.apache.org/r/55677/