Rationalize process wait error checking.
Introduce the WSUCCEEDED() API to test that a forked process exited
successfully. Use it in all the places that were performing this test
bespokely, and update error logs to user WSTRINGIFY() if appropriate.

This also fixes a bug in fetcher where the parent wasn't checking
whether the child was signaled.

Review: https://reviews.apache.org/r/55683/