Fixed unsafe usage of process pointer in loop.hpp.
The `loop(...)` helper spawns a libprocess actor to execute some
lambda (in a loop, of course).  This actor is owned by the
libprocess GC actor, but the body of the `loop` passes a copy of that
pointer into a Future callback.  This will potentially segfault if
the actor terminates outside of the `loop`.

Instead, the `loop` should use the `PID` of the actor.  This change
will make the `loop` code safer, just in case a future commit terminates
the loop actor (such as the use of `process::finalize`).

Review: https://reviews.apache.org/r/55740/