Eliminated a race in libprocess 'Loop'.
In `process::internal::Loop`, we intend to synchronize all
accesses to the `discard` member, since otherwise there would
be a race between setting and calling this lambda. However,
one assignment of the member remains unsynchronized. This
patch synchronizes this assignment to avoid a race.

Review: https://reviews.apache.org/r/55787/