Used `os::host_default_path()` in several locations.
Several of the codepaths that eventually launch a subprocess will check
for a `PATH` environment variable.  If this value is not set, it uses
a default value.

This commit replaces these hard-coded default values with the newly
added stout function `host_default_path`.  The return value of
this function differs depending on the host (and the OS), meaning that
the return value of `host_default_path` may or may not be portable.

Review: https://reviews.apache.org/r/55037/