Updated Resources::apply to handle Resource.AllocationInfo.
Note that previously, `Resource` did not contain the
`AllocationInfo` field.

For backwards compatibility with old schedulers and tooling,
the master injects the allocation info into operations to
ensure that operations can apply correctly to the offered
resources (which are allocated).

When applying these operations with allocated resources to
resources that are unallocated (e.g. applying an operation
to update the agent's total resources), the caller must either
strip the allocation info from the operations prior to applying,
or the apply code needs to ignore the allocation info. For now,
we took the approach of ignoring the allocation info within the
`apply()` function in order to simplify the call-sites.

Note that we assume no `Resources` store a mix of allocated and
unallocated resources, which is brittle and enforcement of this
invariant should be added.

We should consider moving the adjustment code from
`Resources::apply()` up into the call sites. We do this already
via injecting the allocation info into operations, but we don't
have a helper for stripping the allocation info from operations.

Review: https://reviews.apache.org/r/55828