Update the allocator to handle frameworks with multiple roles.
The allocator now sets `Resource.AllocationInfo` when performing
an allocation, and the interface is updated to expose allocations
for each role within a framework.

The allocator also assumes that `Resource.AllocationInfo` is set
for inbound resources that are allocated (e.g. when adding an agent,
when adding a framework, when recovering resources). Note however,
that the necessary changes to the master and agent to enforce this
will be done via separate patches.

Review: https://reviews.apache.org/r/55870