Added `Master::isRemovable(const TaskState&)`.
This determines whether a task in the given state can safely be
discarded from the master's in-memory state. When a task becomes
removable, we move the task from the master's main task data structures
to a fixed-size cache (either the "unreachable" or "completed" task
list, as appropriate).

Review: https://reviews.apache.org/r/55475/