Prevented task launches that reuse unreachable task IDs.
The master keeps an in-memory cache of task IDs that have recently been
marked unreachable. The master now consults this cache to reject task
launch attempts that reuse one of these recently unreachable task IDs
(such tasks are not terminal and may resume running in the future). This
check is not complete (we won't detect all cases in which unreachable
task IDs are reused), but preventing this from happening in the common
case seems worth doing. See MESOS-6785 for details.

Review: https://reviews.apache.org/r/54793/