Changed "master/tasks_unreachable" metric from counter to gauge.
The number of unreachable tasks should be a gauge, because it can
increase and decrease over time. Note that we report the number of
unreachable tasks that the master knows about; this is a subset of the
"true" number of unreachable tasks, because the master only keeps a
limited-size cache of unreachable tasks for each framework.

Review: https://reviews.apache.org/r/55808/