Moved check for 'has_parent()' in docker containerizer destroy.
Ideally, we would do this check as the first thing we do after
entering this function. However, the containerizer API currently
requires callers of `launch()` to also call `destroy()` if the launch
fails (MESOS-6214). As such, putting the check at the top of this
function would cause the containerizer to crash if the launch failure
was due to the container having its `parent` field set. Once we remove
the requirement for `destroy()` to be called explicitly after launch
failures, we should move this check back to the top of this function.

Review: https://reviews.apache.org/r/55778/