Allowed `Master::Framework::updateFrameworkInfo` to return errors.
This allows us to confirm that an updated `FrameworkInfo` is fully
compatible with an already known one. By performing this check in
`updateFrameworkInfo` we can be sure that both the new and old
`FrameworkInfo` are available (e.g., after framework reregistration,
or master or framework failover).

Review: https://reviews.apache.org/r/55571/