Fixed flakiness in TaskValidationTest.TaskReusesUnreachableTaskID.
The previous coding of the test partitioned one agent and then launched
a second agent. However, when run as root, the second agent discovered
(via global cgroups) the task that was started by the first agent, which
it then proceeded to kill.

To avoid this, we now shutdown the first agent after partitioning it and
before starting the second agent.

Review: https://reviews.apache.org/r/55911/