Added TODOs to make the master code safer.
The executor and tasks fields in the master's structs for agents
and frameworks are publicly visible, but yet for correctness it
is required that callers made modifications through the member
functions which carry additional logic (e.g. tracking allocated
resources). The TODO here is to make these private and to provide
public views that are const.

Review: https://reviews.apache.org/r/55968