Updated the master to handle non-MULTI_ROLE schedulers.
When communicating with non-MULTI_ROLE schedulers, the master
must inject the allocation info from the offer in order so that
the operations can be treated as consistently having allocation
info set. This is necessary when applying operations to allocated
resources that contain allocations to multiple roles (e.g. the
sum of allocated resources on an agent).

Review: https://reviews.apache.org/r/55969