Updated master to handle non-MULTI_ROLE agents.
With the addition of MULTI_ROLE framweork support, allocated
resources have a `Resource.AllocationInfo` set. While the new
MULTI_ROLE capable agents will be sending allocated resources,
the old agents may not be since they may not preserve the
unknown fields.

To cope with this, the master ensures the `Resource.AllocationInfo`
is set for non-MULTI_ROLE capable agents, by injecting it. Note
that we can only do this so long as it is unambiguous! This will
be prevented by not allowing frameworks with multiple to use
agents without the MULTI_ROLE support, see: MESOS-6940.

Review: https://reviews.apache.org/r/55972