CMake: Added `GroupSource` function to automate IDE source grouping.
CMake allows users to declare groups of source files, which it uses to
(among other things) present source in a directory-like tree of files in
IDEs like XCode and Visual Studio.

Currently this is a manual process: we group the source in each folder
manually and declare it as a source group to CMake.

This commit will introduce a CMake macro that automates this process
away, providing control over many aspects, such as where the group tree
should be rooted, what the root should be named, and so on.

This macro indirectly partially addresses MESOS-3542, which will help us
to separate out Mesos builds into many libraries, rather than one
single, monolithic libmesos.

Review: https://reviews.apache.org/r/55599/