Fixed a few executor segfaults during cleanup.
The addition of `process::finalize` at the end of some binaries
led to some segfaulting at the end of the binaries' lifetimes.
This is mostly due calling destructors of libprocess actor
wrappers, which require an initialized libprocess to function.

This commit explicitly calls the destructors on the default
and docker executor actors prior to calling `process::finalize`.

Review: https://reviews.apache.org/r/56001