Fixed MULTI_ROLE related bugs when updating framework info.
The first issue is that we need to update the capabilities member
to reflect the new capabilities.

The second issue is that when we allow an upgrade or downgrade
to or from MULTI_ROLE, we need to update the `role` and `roles`
fields of `FrameworkInfo`.

Review: https://reviews.apache.org/r/56004