CMake: Bumped CMake version on Windows, and enforce with check.
This commit will bump the required CMake version on Windows from 3.5.2
(documented requirement, but not enforced) to 3.6.3, and then add a
configure-time check to ensure that this is true before we build.

On Windows, bumping the required version is much less of a big deal than
Unix, mainly because Windows does not have a package manager. The burden
on developers is about the same, independent of the version, because the
most convenient installation option is to use a `.msi`. Additionally,
CMake is not widely used in the community, and where it is, the vast
majority of users are building on Unix, so there are few constraints
imposed by existing CI systems and users.

The primary reason to bump the version, though, is that the CMake
featureset on Windows tends (consciously or not) to straggle a bit
behind the Unix featureset, and recently it has come to our attention
that (due to our ignorance of this issue) we have come to depend on some
subtle features that are available on Unix, but not on Windows, until
CMake 3.6.

Review: https://reviews.apache.org/r/56037/