Disabled ability to launch default executor with ContainerInfo.
The default executor used to launch TaskGroups is generated by the
agent.  The agent generates the executor's `CommandInfo`, hence why
the user may not specify the `CommandInfo` in the `LAUNCH_TASK_GROUP`
call.

This commit adds similar restrictions for `ContainerInfo` plus the
default executor.  The `CommandInfo` constructed by the agent expects
to be run in the same environment as the agent process.  This commit
prevents the user from specifying a `DockerInfo` or a container image
along with the default executor.

If the user explicitly wants to use the default executor, they could
always package the default executor's binary and libraries into a
container and launch it like any other custom executor.

Review: https://reviews.apache.org/r/56340