Fix handling of total resources in role and quote role sorters.
Currently the total resources in role and quota role sorters are tied to
the agent's total resources (which we don't update during allocation
and recovery of resources). To be consistent with this we shouldn't use
the framework's allocation (which could contain additional allocations
for shared resources) to update them in `updateAllocation()`. We
instead use the agent's total resources to update the two sorters'.

Review: https://reviews.apache.org/r/53096/