Use the stout ELF parser to collect Linux rootfs files.
The dependencies for the programs we need in the Linux root
filesystem vary over time and across distributions. Since stout
has support for parsing the library dependencies of ELF binaries,
use this to collect the necessary dependencies when constructing
a root filesystem for the tests.

Review: https://reviews.apache.org/r/53791/