Prevent gaming of the quota system due to demand:quota ratio mismatch.
It is currently possible to game the quota system since the allocator
will perform allocations until the quotas for all resource types are
satisfied. This can lead to allocations for certain resource type
far exceeding the set quota.

In order to prevent this issue, this patch changes the notion of what
constitutes a "satisfied quota". Where before a quota could only be
satisfied when allocations had been met up to the set quota for every
resource type, we now consider a quota satisfied as soon as at least
one of the resource types is allocated up to the set quota. What this
means is that if the allocation and quota vectors are of the same
proportion, this new technique behaves the same as the previous
technique. However, if the allocation and quota vectors have different
resource type ratios, then the quota becomes satisfied as soon as
the any quota guarantee is hit.

Longer term, we can consider allocating the resource types that have
not reached the quota guarantee, but this is non-trivial (consider
the case of having to allocate only memory, this means we likely
should only offer it in such a manner that it allows an existing
container to increase its memory; it cannot be used to create a
new container unless there is a mechanism to obtain best-effort
cpu/disk).

Also note that the quota guarantee is currently applied as a limit,
whereas longer term we would like to (1) make these distinct, and
(2) distinguish quota and fair sharing via non-revocable and
revocable resources, respectively.

Review: https://reviews.apache.org/r/55625/