Fixed incorrect `CHECK_EQ`s in `WindowsFD`.
We store both the `CRT` and `HANDLE` in `WindowsFD` whether we construct
from `CRT` or `HANDLE` by using `_get_osfhandle` and `_open_osfhandle`.

Thus, accessing the `CRT` via `crt()` does not require that it be in the
`CRT` state, but rather `CRT` or `HANDLE`, for example.

Review: https://reviews.apache.org/r/56354