Introduced a `Container` struct on the default executor.
This helps us in consolidating various metadata associatd with
an active child container. Also, this would be used in the future
for figuring out the task group corresponding to a child container
and then killing the other child containers belonging to the task
group when one of the tasks fail; or if a task in the task group
is explicitly killed by the scheduler.

Review: https://reviews.apache.org/r/56269/