Fixed a regression in the execute CLI command.
With the addition of `Resource.allocation_info` to support MULTI_ROLE
frameworks, the execute helper has pulled in the updated `Resources`
library and needs to be updated. We strip the allocation info to
ensure that `execute` will work against both old and new masters.

Review: https://reviews.apache.org/r/56408