Added platform-independent constants `DEV_NULL` and `TRUE_COMMAND`.
The `DEV_NULL` constant points to `/dev/null` on Posix and `NUL` on
Windows.  In both platforms, it can be used to redirect output
into digital oblivion.

The `TRUE_COMMAND` is a test constant that equals the `true` shell
command on Posix.  On Windows, we emulate this command by invoking
a shell with an inline command to exit with a successful exit code.

Review: https://reviews.apache.org/r/55546/