Windows: Added health checker to build.
This removes the Windows guard around the health checker CMake build,
thereby enabling the build of `mesos-tcp-connect` on Windows.

In addition, this replaces some instances of "/dev/null" strings with
the newly added constant `os::DEV_NULL`.  This is necessary for the
health checker code to run successfully on Windows.

Review: https://reviews.apache.org/r/55549/