Windows: Enabled health checker tests.
This ports all the health check tests running on Posix to also run
on Windows, with the exception of two tests that depend on curl binary,
which is not on the PATH on Windows.

Porting the tests themselves involve replacing some command strings
with macros or constants and writing a PowerShell version of the
health check command (one that fails every other invocation).

Review: https://reviews.apache.org/r/55550/