Added CMake to standard documentation.
This qualifies the existing "configure" options with "autotools"
to differentiate them from CMake build options.  And this adds
CMake build options to a separate document.  This new document
may be extended in future to document more of the features,
practices, or nuances of the CMake build.

Review: https://reviews.apache.org/r/55749/